+++ source/components/com_asinustimetracking/admin/_update/script.postprocess.php
@@ -1,8 +1,5 @@
 <?php
 /**
- * @package      pkg_projectfork
- * @subpackage   com_projectfork
- *
  * @author       Tobias Kuhn (eaxs)
  * @copyright    Copyright (C) 2006-2013 Tobias Kuhn. All rights reserved.
  * @license      http://www.gnu.org/licenses/gpl.html GNU/GPL, see LICENSE.txt