+++ source/components/com_asinustimetracking/site/views/monthlyreport/tmpl/default_excel-template-1.php
@@ -1,7 +1,19 @@
 <?php
+/**
+ * @package		Joomla.Site
+ * @subpackage	com_asinustimetracking
+ * @copyright	Copyright (c) 2014, Valentin Despa. All rights reserved.
+ * @author		Valentin Despa - info@vdespa.de
+ * @link		http://www.vdespa.de
+ * @license 	GNU General Public License version 3. See LICENSE.txt or http://www.gnu.org/licenses/gpl-3.0.html
+ */
+
+defined('_JEXEC') or die;
+
 // Try to increase memory limit
 ini_set('memory_limit', '128M');
 
+// FIXME, this should be be hardcoded
 $f = JPATH_ROOT . '/media/com_asinustimetracking/report-templates/monthly-report-template-1.xlsx';
 
 /** Load template to a PHPExcel Object  **/